Refactor: user 패키지 -> sub.user 패키지로 이동
