+++ src/main/java/com/sppart/admin/sub/user/dto/CurrentUser.java
@@ -1,14 +1,11 @@
-package com.sppart.admin.user.interceptor;
+package com.sppart.admin.sub.user.dto;
 
-import com.sppart.admin.user.domain.Role;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+@Target({ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Auth {
-
-    Role[] roles();
+public @interface CurrentUser {
 }
