+++ src/main/java/com/sppart/admin/sub/user/interceptor/Auth.java
@@ -1,11 +1,14 @@
-package com.sppart.admin.user.dto;
+package com.sppart.admin.sub.user.interceptor;
 
+import com.sppart.admin.sub.user.domain.Role;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface CurrentUser {
+@Target(ElementType.METHOD)
+public @interface Auth {
+
+    Role[] roles();
 }
