+++ src/main/resources/mappers/sub/UserMapper.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.sppart.admin.user.domain.mapper.UserMapper">
+<mapper namespace="com.sppart.admin.sub.user.domain.mapper.UserMapper">
 
-    <select id="findById" parameterType="String" resultType="com.sppart.admin.user.domain.Users">
+    <select id="findById" parameterType="String" resultType="com.sppart.admin.sub.user.domain.Users">
         SELECT u.admin_user_id, u.id, u.password, r.name
         FROM `admin_user` u
                  INNER JOIN role r ON u.role_id = r.id
